refactor(calculator): separate create and use
