+++ src/test/java/calculator/CalculatorTest.java
@@ -69,6 +69,6 @@ public class CalculatorTest {
     }
 
     private Calculator<Integer, String> createCalculator() {
-        return new StringAddCalculator();
+        return new StringAddCalculator(new PartsGenerator());
     }
 }
